requirements.txt
# Core dependencies (latest versions compatible with Python 3.12)
langchain>=0.3.0
langgraph>=0.2.0
langchain-openai>=0.2.0
langchain-anthropic>=0.3.0
langchain-google-genai>=2.0.0
langchain-core>=0.3.0

# HTTP client
httpx>=0.27.0
requests>=2.32.0

# Web framework for observability
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=13.0

# Data handling
pydantic>=2.10.0
python-dotenv>=1.0.1

# Logging and monitoring
structlog>=24.4.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
